+++ x2engine/protected/config/X2Config.php
@@ -5,8 +5,8 @@ $host = 'localhost';
 $user = 'root';
 $pass = '';
 $dbname = 'x2engine';
-$version = '3.7.3b';
-$buildDate = 1392445087;
-$updaterVersion = '3.7.3b';
+$version = '3.7.3';
+$buildDate = 1392772447;
+$updaterVersion = '3.7.3';
 $language='en';
 ?>