+++ x2engine/protected/tests/unit/components/permissions/X2PermissionsBehaviorTest.php
@@ -18,13 +18,13 @@ class X2PermissionsBehaviorTest extends X2TestCase {
     
     public function testGetAccessConditions() {
         $behavior = new X2PermissionsBehavior;
-        $this->assertEquals(true, is_array($behavior->getAccessConditions(0, 0, 'admin')));
+        $this->assertEquals(true, is_array($behavior->getAccessConditions(0, false , 'admin')));
         
-        $accessLevel3 = $behavior->getAccessConditions(3, 1, 'admin');
+        $accessLevel3 = $behavior->getAccessConditions(3, 'visibility', 'admin');
         $this->assertEquals('TRUE', $accessLevel3[0]['condition']);
         $this->assertEquals('AND', $accessLevel3[0]['operator']);
         
-        $accessLevel2 = $behavior->getAccessConditions(1, 1, 'admin');
+        $accessLevel2 = $behavior->getAccessConditions(1, 'visibility', 'admin');
         $this->assertEquals('t.assignedTo="admin"', $accessLevel2[0]['condition']);
         $this->assertEquals('OR', $accessLevel2[0]['operator']);